Refactor some imports
