+++ src/api/reductions/float_arithmetic.rs
@@ -16,7 +16,7 @@ macro_rules! impl_reduction_float_arithmetic {
             #[inline]
             pub fn sum(self) -> $elem_ty {
                 #[cfg(not(target_arch = "aarch64"))] {
-                    use super::llvm::simd_reduce_add_ordered;
+                    use crate::llvm::simd_reduce_add_ordered;
                     unsafe { simd_reduce_add_ordered(self.0, 0 as $elem_ty) }
                 }
                 #[cfg(target_arch = "aarch64")] {
@@ -43,7 +43,7 @@ macro_rules! impl_reduction_float_arithmetic {
             #[inline]
             pub fn product(self) -> $elem_ty {
                 #[cfg(not(target_arch = "aarch64"))] {
-                    use super::llvm::simd_reduce_mul_ordered;
+                    use crate::llvm::simd_reduce_mul_ordered;
                     unsafe { simd_reduce_mul_ordered(self.0, 1 as $elem_ty) }
                 }
                 #[cfg(target_arch = "aarch64")] {
