+++ src/codegen/math/float/sqrte.rs
@@ -3,8 +3,8 @@
 
 // FIXME 64-bit 1 elem vectors sqrte
 
-use super::super::super::llvm::simd_fsqrt;
 use crate::*;
+use crate::llvm::simd_fsqrt;
 
 crate trait Sqrte {
     fn sqrte(self) -> Self;
